Unify dates in the store
