+++ src/stores/playbackStore.ts
@@ -2,7 +2,6 @@ import { Metric } from 'anomaly-detection';
 import { defineStore, acceptHMRUpdate } from 'pinia';
 import { metricsApi } from '../services/apiService';
 import { formattedDate, getDatesBetween, subtractDays } from '../utils/date';
-import { useUIStore } from './uiStore';
 import { useRegionDetailedStore } from './regionDetailedStore';
 
 export const usePlaybackStore = defineStore('playbackStore', {
@@ -45,7 +44,6 @@ export const usePlaybackStore = defineStore('playbackStore', {
       return response.data;
     },
     togglePlayback() {
-      const uiStore = useUIStore();
       const regionDetailedStore = useRegionDetailedStore();
       this.playbackEnabled = !this.playbackEnabled;
       if (this.playbackEnabled) {
@@ -55,9 +53,6 @@ export const usePlaybackStore = defineStore('playbackStore', {
         this.playbackCurrentIndex = 0;
         this.data = null; // Clear previous data
         regionDetailedStore.$reset(); // Reset region detailed store
-      } else {
-        // Disabling playback
-        uiStore.fetchLastDate();
       }
     },
     updateCurrentDate(index: number) {
