+++ src/stores/uiStore.ts
@@ -1,12 +1,11 @@
 import { defineStore, acceptHMRUpdate } from 'pinia';
-import { metricsApi } from '../services/apiService';
 import routesNames from '../router/routesNames';
 import { formattedDate } from '../utils/date';
+import { useMapStore } from './mapStore';
+import { usePlaybackStore } from './playbackStore';
 
 export const useUIStore = defineStore('uiStore', {
   state: () => ({
-    date: '2025-01-01',
-    fetchingDate: true,
     appWidth: window.innerWidth as number,
     regionDetailDrawerWidth: 0,
     currentTab: routesNames.anomalyMap,
@@ -14,30 +13,15 @@ export const useUIStore = defineStore('uiStore', {
 
   getters: {
     formattedDate: (state) => {
-      return formattedDate(state.date);
+      const mapStore = useMapStore();
+      const playbackStore = usePlaybackStore();
+      return playbackStore.playbackEnabled
+        ? formattedDate(playbackStore.playbackCurrentDate)
+        : formattedDate(mapStore.currentDate);
     },
   },
 
-  actions: {
-    // * Date
-    async fetchLastDate() {
-      try {
-        this.fetchingDate = true;
-        const response = await metricsApi.lastDateRetrieve();
-        if (response.status === 200 && response.data) {
-          this.setDate(response.data.date);
-          this.fetchingDate = false;
-        } else {
-          throw new Error('Failed to fetch last date');
-        }
-      } catch (error) {
-        console.error('Error fetching last date:', error);
-      }
-    },
-    setDate(newDate: string) {
-      this.date = newDate;
-    },
-  },
+  actions: {},
 });
 
 if (import.meta.hot) {
